- актуализировал в соответсвии с серверной версией проекта. - подготовил все для переезда JPA - кастомный контроль транзакций перееха в SampleRepository - Examination и Channel пишутся теперь по команде а не гибирнейтом - перееха с @ClassID на @EmbeddedId в классах Channel и Sample - причесал все entity - актуализировал ExaminationServiceTest - рефактеринг

+++ src/main/java/ru/gsa/biointerface/domain/entity/PatientRecord.java
@@ -21,7 +21,7 @@ public class PatientRecord implements Serializable, Comparable<PatientRecord> {
     @NotNull(message = "Id can't be null")
     @Min(value = 1, message = "Id can't be lass then 1")
     @Id
-    private long id;
+    private int id;
 
     @NotNull(message = "Second name can't be null")
     @NotBlank(message = "Second name can't be blank")
@@ -62,7 +62,7 @@ public class PatientRecord implements Serializable, Comparable<PatientRecord> {
     public PatientRecord() {
     }
 
-    public PatientRecord(long id, String secondName, String firstName, String patronymic, Calendar birthday, Icd icd, String comment) {
+    public PatientRecord(int id, String secondName, String firstName, String patronymic, Calendar birthday, Icd icd, String comment) {
         this.id = id;
         this.secondName = secondName;
         this.firstName = firstName;
@@ -73,11 +73,11 @@ public class PatientRecord implements Serializable, Comparable<PatientRecord> {
         this.icd = icd;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
